requirements.txt
pandas==2.2.2
pyarrow==15.0.2
fastparquet==2023.10.1
numpy==1.26.4
matplotlib==3.8.4
seaborn==0.13.2
scipy==1.13.0
scikit-learn==1.4.2
requests==2.31.0
python-dateutil==2.9.0
tqdm==4.66.4
colorlog==6.8.2